requirements.txt
flask==1.1.2
biopython==1.78
gunicorn==20.0.4
Werkzeug==0.16.0
wtforms==2.3.3
flask_bootstrap==3.3.7.1
Flask-WTF == 0.14.3
Flask-Uploads == 0.2.1
XlsxWriter==1.3.7
Flask-Login == 0.5.0 
Flask-Table == 0.5.0
